Adding player health, taking damage, and UI reactiveness with a widget based system.
In this commit, I added a new health system in for the player to use and tied it in to my first enemy type's heavy attack for testing. I then created a UI manager + widget system that will allow me to store and handle different UI widgets for things like health, weapons, buffs, etc. So far I have implemented only the health widget, which now has a working animated health bar and health number as well.